Reorganize namespaces
